# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ExtractionJobCreated {
  job_id: string;
  message: string;
}

export type ExtractionJobStatus =
  (typeof ExtractionJobStatus)[keyof typeof ExtractionJobStatus];

export const ExtractionJobStatus = {
  pending: "pending",
  running: "running",
  complete: "complete",
  failed: "failed",
} as const;

export type ExtractionJobStage =
  (typeof ExtractionJobStage)[keyof typeof ExtractionJobStage];

export const ExtractionJobStage = {
  reading_pdf: "reading_pdf",
  running_ner: "running_ner",
  analyzing_deal: "analyzing_deal",
  extracting_fields: "extracting_fields",
  validating: "validating",
  formatting: "formatting",
  complete: "complete",
  failed: "failed",
} as const;

export type ExtractionDetailDealStructure = { [key: string]: unknown } | null;

export type ExtractionDetailFormattedData = { [key: string]: string };

export type ExtractionDetailRawData = { [key: string]: unknown } | null;

export interface ExtractionDetail {
  id: number;
  terms_filename: string;
  credit_memo_filename?: string | null;
  deal_structure?: ExtractionDetailDealStructure;
  formatted_data: ExtractionDetailFormattedData;
  raw_data?: ExtractionDetailRawData;
  ner_warnings: string[];
  fields_populated: number;
  fields_total: number;
  completion_pct: number;
  created_at: string;
}

export interface ExtractionJob {
  job_id: string;
  status: ExtractionJobStatus;
  stage: ExtractionJobStage;
  stage_label: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  error?: string | null;
  extraction_id?: number | null;
  result?: ExtractionDetail | null;
}

export interface ExtractionSummary {
  id: number;
  terms_filename: string;
  credit_memo_filename?: string | null;
  deal_type?: string | null;
  loan_program?: string | null;
  borrower_name?: string | null;
  loan_amount?: string | null;
  fields_populated: number;
  fields_total: number;
  completion_pct: number;
  created_at: string;
  has_ner_warnings: boolean;
}

export interface ExtractionList {
  extractions: ExtractionSummary[];
  total: number;
  page: number;
  per_page: number;
  total_pages: number;
}

export interface SuccessResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
}

export type StartExtractionBody = {
  /** Terms & Conditions PDF */
  terms_pdf: Blob;
  /** Optional Credit Memo PDF */
  credit_memo_pdf?: Blob;
};

export type ListExtractionsParams = {
  page?: number;
  per_page?: number;
};

export type DownloadExtraction200 = { [key: string]: unknown };