# lib/api-zod/src/generated/types/extractionDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractionDetailDealStructure } from "./extractionDetailDealStructure";
import type { ExtractionDetailFormattedData } from "./extractionDetailFormattedData";
import type { ExtractionDetailRawData } from "./extractionDetailRawData";

export interface ExtractionDetail {
  id: number;
  terms_filename: string;
  credit_memo_filename?: string | null;
  deal_structure?: ExtractionDetailDealStructure;
  formatted_data: ExtractionDetailFormattedData;
  raw_data?: ExtractionDetailRawData;
  ner_warnings: string[];
  fields_populated: number;
  fields_total: number;
  completion_pct: number;
  created_at: string;
}