# lib/api-zod/src/generated/types/extractionJob.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractionDetail } from "./extractionDetail";
import type { ExtractionJobStage } from "./extractionJobStage";
import type { ExtractionJobStatus } from "./extractionJobStatus";

export interface ExtractionJob {
  job_id: string;
  status: ExtractionJobStatus;
  stage: ExtractionJobStage;
  stage_label: string;
  /**
   * @minimum 0
   * @maximum 100
   */
  progress: number;
  error?: string | null;
  extraction_id?: number | null;
  result?: ExtractionDetail | null;
}