# lib/api-zod/src/generated/types/extractionJobCreated.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractionJobCreated {
  job_id: string;
  message: string;
}