# lib/api-zod/src/generated/types/extractionList.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractionSummary } from "./extractionSummary";

export interface ExtractionList {
  extractions: ExtractionSummary[];
  total: number;
  page: number;
  per_page: number;
  total_pages: number;
}