# lib/api-zod/src/generated/types/extractionSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * SBA Loan Data Extraction Tool API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractionSummary {
  id: number;
  terms_filename: string;
  credit_memo_filename?: string | null;
  deal_type?: string | null;
  loan_program?: string | null;
  borrower_name?: string | null;
  loan_amount?: string | null;
  fields_populated: number;
  fields_total: number;
  completion_pct: number;
  created_at: string;
  has_ner_warnings: boolean;
}